pyproject.toml
[project]
name = "test-docx"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "python-docx>=1.2.0",
]
